Develop

+++ src/main/java/com/rest/springboot/controller/EmployeeController.java
@@ -31,9 +31,8 @@ public class EmployeeController {
 
 	@GetMapping("/{id}")
 	public ResponseEntity<EmployeeDetails> getEmployee(@PathVariable("id") Long id) {
-		System.out.println(id);
-		EmployeeDetails employee = employeeService.getEmployee(id);
-		if (employee != null) {
+		if (employeeService.isExist(id)) {
+			EmployeeDetails employee = employeeService.getEmployee(id);
 			return new ResponseEntity<EmployeeDetails>(employee, HttpStatus.OK);
 		} else {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -46,14 +45,21 @@ public class EmployeeController {
 	}
 
 	@PutMapping("/{id}")
-	public EmployeeDetails updateEmployeeDetails(@PathVariable("id") Long id,
+	public ResponseEntity<EmployeeDetails> updateEmployeeDetails(@PathVariable("id") Long id,
 			@RequestBody EmployeeDetails employeeDetail) throws Exception {
-		return employeeService.updateEmployeeDetails(id, employeeDetail);
+		if (employeeService.isExist(id)) {
+			return new ResponseEntity<>(employeeService.updateEmployeeDetails(id, employeeDetail), HttpStatus.OK);
+		}
+		return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 	}
 
 	@DeleteMapping("/{id}")
-	public void deleteEmployeeDetail(@PathVariable("id") Long id) {
-		employeeService.deleteEmployeeDetails(id);
+	public ResponseEntity<Void> deleteEmployeeDetail(@PathVariable("id") Long id) {
+		if (employeeService.isExist(id)) {
+			employeeService.deleteEmployeeDetails(id);
+			return new ResponseEntity<Void>(HttpStatus.OK);
+		}
+		return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
 	}
 
 }
